Add fixes
- fix invalid download step
- use prefixed product IDs on ratings import job
- downgrade log errors to warnings to reduce noise
- fix product controller to handle missing products
- make product middleware reusable
