+++ cartridges/bm_bazaarvoice/cartridge/scripts/jobs/bvRatingImport.js
@@ -8,12 +8,14 @@ var Status = require('dw/system/Status');
 var XMLStreamConstants = require('dw/io/XMLStreamConstants');
 var XMLStreamReader = require('dw/io/XMLStreamReader');
 var Logger = require('dw/system/Logger').getLogger('Bazaarvoice', 'bvRatingImport.js');
+var Site = require('dw/system/Site');
 
-var BVHelper = require('*/cartridge/scripts/lib/libBazaarvoice').getBazaarVoiceHelper();
+var BVHelper = require('bm_bazaarvoice/cartridge/scripts/lib/libBazaarvoice').getBazaarVoiceHelper();
 var localeHelper = require('./util/localeHelper');
 
 module.exports.execute = function() {
     try {
+        var siteId = Site.getCurrent().getID();
         //generate a locale map
         //also generate a reverse map of:
         //   BV locale ===> Array of DW locales
@@ -24,18 +26,18 @@ module.exports.execute = function() {
         if(multiLocale) {
             bvLocaleMap = localeHelper.getBVLocaleMap(localeMap);
         }
-        
+
         //get the rating feed we just downloaded
         var tempPath = [File.TEMP, 'bv', 'ratings', 'ratings.xml'].join(File.SEPARATOR);
         var tempFile = new File(tempPath);
         if(!tempFile.exists()) {
             throw new Error('TEMP/bv/ratings/ratings.xml does not exist!');
         }
-        
+
         //open the feed and start stream reading
         var fileReader = new FileReader(tempFile, 'UTF-8');
         var xmlReader = new XMLStreamReader(fileReader);
-        
+
         while(xmlReader.hasNext()) {
             xmlReader.next();
             if (xmlReader.getEventType() === XMLStreamConstants.START_ELEMENT && xmlReader.getLocalName() === 'Product') {
@@ -45,18 +47,34 @@ module.exports.execute = function() {
                 var bvRatingRange = '';
                 var ns = productXML.namespace();
                 var id = productXML.ns::ExternalId.toString();
-                var product = id ? ProductMgr.getProduct(BVHelper.decodeId(id)) : null;
+                var prefix = siteId + '-';
+                var product = null;
+
+                if (id.indexOf(prefix) > -1) {
+                    id = id.replace(prefix, '');
+
+                    var reviewProduct = ProductMgr.getProduct(BVHelper.decodeId(id));
+
+                    if (reviewProduct) {
+                        if (reviewProduct.isVariant()) {
+                            product = reviewProduct.getMasterProduct();
+                        } else {
+                            product = reviewProduct;
+                        }
+                    }
+                }
+
                 if(product) {
                     if(multiLocale) {
-                        
+
                         var localeItemList = productXML.ns::ReviewStatistics.ns::LocaleDistribution.ns::LocaleDistributionItem;
                         if(localeItemList && localeItemList.length() > 0) {
                             for(var i = 0; i < localeItemList.length(); i++) {
                                 var localeItem = localeItemList[i];
-                                
+
                                 var bvLocale = localeItem.ns::DisplayLocale.toString();
                                 if(bvLocale) {
-                                    
+
                                     if (!empty(localeItem.ns::ReviewStatistics.ns::AverageOverallRating.toString())) {
                                         bvAverageRating = localeItem.ns::ReviewStatistics.ns::AverageOverallRating.toString();
                                     }
@@ -66,7 +84,7 @@ module.exports.execute = function() {
                                     if (!empty(localeItem.ns::ReviewStatistics.ns::OverallRatingRange.toString())) {
                                         bvRatingRange = localeItem.ns::ReviewStatistics.ns::OverallRatingRange.toString();
                                     }
-                                    
+
                                     var dwLocales = bvLocaleMap.get(bvLocale);
                                     if(dwLocales && dwLocales != null){
                                         for(var j = 0; j < dwLocales.length; j++) {
@@ -80,9 +98,9 @@ module.exports.execute = function() {
                                 }
                             }
                         }
-                        
+
                     } else {
-                        
+
                         if (!empty(productXML.ns::ReviewStatistics.ns::AverageOverallRating.toString())) {
                             bvAverageRating = productXML.ns::ReviewStatistics.ns::AverageOverallRating.toString();
                         }
@@ -92,7 +110,7 @@ module.exports.execute = function() {
                         if (!empty(productXML.ns::ReviewStatistics.ns::OverallRatingRange.toString())) {
                             bvRatingRange = productXML.ns::ReviewStatistics.ns::OverallRatingRange.toString();
                         }
-                        
+
                         product.custom.bvAverageRating = bvAverageRating;
                         product.custom.bvReviewCount = bvReviewCount;
                         product.custom.bvRatingRange = bvRatingRange;
@@ -100,14 +118,14 @@ module.exports.execute = function() {
                 }
             }
         }
-        
+
         xmlReader.close();
         fileReader.close();
-        
+
     } catch(e) {
         Logger.error('Exception caught: ' + e.message);
         return new Status(Status.ERROR, 'ERROR', e.message);
     }
-    
+
     return new Status(Status.OK);
 };
