+++ cartridges/int_bazaarvoice_sfra/cartridge/templates/default/bv/display/pdpsummary.isml
@@ -1,12 +1,24 @@
 <!--- TEMPLATENAME: pdpSummary.isml --->
- <isscript>
-    var assets = require('*/cartridge/scripts/assets');
-    assets.addJs('/js/pdpSummary.js');
-</isscript>
 <isif condition="${pdict.bvDisplay.showSummary}">
     <div data-bv-show="rating_summary" data-bv-product-id="${pdict.bvDisplay.bvPid}"  id="data-bv-show" class="bv-hidden">
         <isif condition="${pdict.bvDisplay.rr.seo && pdict.bvDisplay.rr.seo.aggregateRating}">
             <isprint value="${pdict.bvDisplay.rr.seo.aggregateRating}" encoding="off"/>
         </isif>
     </div>
+
+    <script type="text/javascript" language="javascript">
+        window.bvCallback = function (BV) {
+            if (typeof BV !== 'undefined') {
+                BV.reviews.on('show', function () {
+                    // This line assumes SFRA collapsible panels and is called when clicking on the review summary
+                    $('.row.reviews:not(.active)').toggleClass('active');
+                });
+                BV.questions.on('show', function () {
+                    // This line assumes SFRA collapsible panels and is called when clicking on the Q&A summary
+                    $('.row.questions:not(.active)').toggleClass('active');
+                });
+            }
+        };
+    </script>
+    <!-- End Reviews Summary -->
 </isif>
